refactor:/types폴더 구조 변경
