+++ src/components/home/ExhibitionCard.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link';
 import { Calendar, CheckCircle, Clock } from 'lucide-react';
 import { cn } from '@/lib/utils';
 import { formatDate, getStatus } from '@/lib/exhibition/dateStatus';
-import { ExhibitionListItem } from '@/types/exhibitionList';
+import { ExhibitionListItem } from '@/types/exhibition';
 import LikeButton from './LikeButton';
 
 interface ExhibitionCardProps {
