+++ src/components/my-wishlist/WishlistScreen.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react';
 import ArtworkModal from '@/components/my-artworks/ArtworkModal';
 import FilterTab from '@/components/my-artworks/FilterTab';
-import type { Artwork, FilterType } from '@/components/my-artworks/Types';
+import type { Artwork, FilterType } from '@/types/artwork';
 import WishlistCard from './WishlistCard';
 
 interface WishlistScreenProps {
