+++ src/types/artwork.ts
