+++ src/types/exhibition.ts
